Major updates to character editor and testing area

+++ character-monster/actions/upsertThirdPartyApiKeySA.ts
@@ -5,7 +5,7 @@ import { createClient } from '@/utils/supabase/server'
 import {
   ThirdPartyAPIKey,
   ThirdPartyAPIKeySchema,
-  UUIDSchema,
+  SupportedServicesSchema,
 } from '@/lib/schemas'
 import { THIRD_PARTY_KEYS_TABLE } from '@/lib/constants'
 import { cookies, headers } from 'next/headers'
@@ -17,33 +17,37 @@ type ReturnType = {
 
 const inputSchema = z.object({
   apiKey: z.string(),
-  userId: UUIDSchema,
+  service: SupportedServicesSchema,
 })
 
 type InputType = z.infer<typeof inputSchema>
 
-const upsertElevenLabsApiKeySA = async ({
+export const upsertThirdPartyApiKeySA = async ({
   apiKey,
-  userId,
+  service,
 }: InputType): Promise<ServerActionReturn<ReturnType>> => {
   const cookieJar = cookies()
   const headersList = headers()
   const jwt = headersList.get('user-allowed-session')
   const supabase = jwt ? createClient(cookieJar, jwt) : createClient(cookieJar)
   try {
+    // Get user Id
+    const userId = (await supabase.auth.getSession()).data.session?.user.id
+    if (!userId) throw new Error('User ID not found in session')
+
     const newAPIKey: ThirdPartyAPIKey = ThirdPartyAPIKeySchema.parse({
       api_key: apiKey,
       owner: userId,
-      type: 'elevenlabs',
-      endpoint: null,
+      type: service, // Use the service parameter here
+      endpoint: null, // This can be modified or removed as per requirements
     })
 
-    // Delete any exisiting api keys with elevenlabs type
+    // Delete any existing api keys with the provided service type
     const { error: deleteError } = await supabase
       .from(THIRD_PARTY_KEYS_TABLE)
       .delete()
       .eq('owner', userId)
-      .eq('type', 'elevenlabs')
+      .eq('type', service)
 
     const { error } = await supabase
       .from(THIRD_PARTY_KEYS_TABLE)
@@ -59,10 +63,10 @@ const upsertElevenLabsApiKeySA = async ({
     return {
       error: extractErrorMessage(
         error,
-        'Unknown error from upsertElevenLabsApiKeySA.'
+        `Unknown error from upsert${
+          service.charAt(0).toUpperCase() + service.slice(1)
+        }ApiKeySA.`
       ),
     }
   }
 }
-
-export default upsertElevenLabsApiKeySA
